여행 확인하기 완료

+++ src/main/java/com/green/project_quadruaple/trip/model/res/TripDetailRes.java
@@ -17,7 +17,7 @@ public class TripDetailRes extends ResultResponse {
         super(ResponseCode.OK.getCode());
     }
 
-    private double totalDistance;
+    private long totalDistance;
     private int totalDuration;
     private int scheduleCnt;
     private int memoCnt;
